cedar: STP remote loader transfers all Basic.Loadees; blocker is now the unmatched IWDC
Guest login now completes the full remote-loader sequence: Basic.Loadees
plus all 33 BCDs (34 payload files) transfer from chm/cedar/stp-root via
the in-process STP/BSP server (ethernet.c), following the original
STPServerImpl.mesa!9 / STPImpl.mesa!5 state machines (HereIsPList, client
Yes, HereIsFile, data under the 31-Pup window, Yes + Mark EOC). Per-sector
IOCB progression in the PDI bridge unblocks the client's local cache
writes at the 32-page boundary.

The loader then reaches InterpreterTool.bcd and dies with Pilot's Wakeup
Disable Counter climbing without matching DWDC (WDC=0333B, NWW=100000B,
CurrentPSB=0, frozen CurrentTime); PilotMesaProcess.mc BusyWait rejects
idle nonzero WDC. Next step is tracing the original Requeue path, not a
host-side scheduler fix (the WDC-clear and ready-queue re-insert probes
both fail validation and stay diagnostics-only).

Supporting changes:
- cpu: 64-bit cycle counter (long Cedar runs no longer wrap)
- machine: Pilot timer channel mask 0100000B per PilotMesaProcess.mc
- diagnostics: DORADO_FTP_TRACE, DORADO_PDI_BCD_TRACE, DORADO_RM_WATCH/
  _VALUE, DORADO_STORE_TRACE_VA, DORADO_TRACE_GATE, opt-in
  DORADO_PILOT_WDC_RECOVER / DORADO_PILOT_READY_RECOVER probes
- media: chm/cedar 6.1 sources + stp-root release tree,
  CedarDorado-bestof.pdi.gz + matching login snapshot,
  docs/cedar-media-provenance.md, tools/fetch_cedar_loadees.py
- gitignore: tools/rusty-backup/ (external 9 GB tool clone), stray
  dorado/*.pdi scratch images

+++ dorado/include/cpu.h
@@ -417,7 +417,12 @@ typedef struct {
     /* Termination / debug. */
     int      halted;            /* nonzero = stop stepping */
     int      halt_reason;
-    int      cycles;            /* cycles executed */
+    /* The standalone Cedar loader runs for billions of microinstructions.
+     * Keep this distinct from the BaseBoard's clock but give it the same
+     * range; an int wrapped at 2^31 and poisoned deferred-Md timing.  This
+     * remains snapshot-layout compatible with the former int plus alignment
+     * padding before ifu_dispatch_count. */
+    uint64_t cycles;            /* cycles executed */
 
     /* Count of successful IFU opcode dispatches (each real IFUJump that
      * reads an opcode and dispatches to its handler). Incremented in
@@ -470,7 +475,7 @@ void dorado_cpu_init(dorado_cpu *cpu, const dorado_microcode *mc,
 int dorado_cpu_step(dorado_cpu *cpu);
 
 /* Run until halt or `max_cycles` reached. Returns the halt reason. */
-cpu_halt_reason dorado_cpu_run(dorado_cpu *cpu, int max_cycles);
+cpu_halt_reason dorado_cpu_run(dorado_cpu *cpu, uint64_t max_cycles);
 
 /* Enable line-by-line tracing to stderr (or a user FILE*). */
 void dorado_cpu_trace(dorado_cpu *cpu, void *fp);
